+++ models.py
@@ -14,7 +14,6 @@ import kivy
 kivy.require('1.8.0')
 from kivy import platform
 
-
 def get_database():
     if platform == 'android':
         file = '/sdcard/pocketphilosopher/aphorisms.db'
@@ -67,13 +66,39 @@ class Aphorism(BaseModel):
         return "{0}".format(
             (string[:length] + '"...') if len(string) > length else string)
 
+    def RemoveDuplicates(self):
+        """Remove duplicate aphorisms by date created"""
+        dupes = self.raw("SELECT id, aphorism "
+                        "FROM aphorism "
+                        "WHERE aphorism IN (SELECT aphorism "
+                        "                   FROM aphorism "
+                        "                   GROUP BY aphorism "
+                        "                   HAVING COUNT(aphorism) > 1"
+                        "                   ORDER BY aphorism) "
+                        "ORDER BY aphorism, created")
+        last = None
+        removed = 0
+        for a in dupes.execute():
+            if last is None:
+                last = a
+                continue
+            if last.aphorism == a.aphorism:
+                try:
+                    a.delete_instance()
+                except Exception as e:
+                    pass
+                else:
+                    removed = removed + 1
+            else:
+                last = a
+        return removed
 
 def CreateTables():
     """Create database tables for Aphorisms in SQLite"""
     try:
         Aphorism.create_table()
-    except OperationalError:
-        click.secho('Aphorism table already exists!', fg='red')
+    except OperationalError as e:
+        click.secho('Aphorism table already exists!' + e, fg='red')
     else:
         click.secho('Aphorism table successfully created.', fg='green')